modify sth

+++ projects/bingdou-core/src/main/java/com/bingdou/core/mapper/user/IExpGradeMapper.java
@@ -8,7 +8,7 @@ import org.apache.ibatis.annotations.Param;
  */
 public interface IExpGradeMapper {
 
-    Float getExpAmountByLevelId(int levelId);
+    Integer getExpAmountByLevelId(int levelId);
 
     UserExpGrade getUserExpGradeInfo(int userId);
